Create separate file for my_enumarable
