+++ my_enumerable.rb
@@ -1,0 +1,23 @@
+module MyEnumerable
+  def pintar
+    each { |n| print n }
+  end
+
+  def all?(&block)
+    result = true
+    each { |n| result = false unless block.call(n) }
+    puts result
+  end
+
+  def any?(&block)
+    result = false
+    each { |n| result = true if block.call(n) }
+    puts result
+  end
+
+  def filter(&block)
+    result = []
+    each { |n| result << n if block.call(n) }
+    puts result.inspect
+  end
+end
